Add desktop notifications
This adds desktop notifications for Gone to be able to see when a
session has ended. The change introduces a message that will say for
instance that "Your working session has ended, time for a short break".
A Notifier interface has been added to make sure that the feature can be
extended with custom notifiers.

Desktop notifications is nice to have to be able to get an interruption
when a session is done.